define a config value to allow errors to be raised
By defining a new config value that flask_pydantic looks for
(FLASK_PYDANTIC_VALIDATION_ERROR_RAISE), we can support raising an
exception when we have a validation error. If that is set and the
exception is raised, by default, Flask will respond with a 500 error.
However, by coupling that feature with `Flask.register_error_handler`,
the user can catch the thrown exception and return whatever custom error
response the user wishes. Fixes #45.