Detect lists via type annotations
Pydantic no longer has an .is_complex() method. But the meaning of the
code seems to have been to detect lists so that queries like

    a=1&a=2&a=3

can be unflattened to

    a=['1', '2', '3']

when a is List[str]. This implementation should work for both List[...]
and Optional[List[...]] but requires a small dependency on older versions
of Python.